Make `KernelDescriptor` bind to the current CUDA context.
A `KernelDescriptor` will now store a reference to the current CUDA context
and only compare equal to another `KernelDescriptor` if they have a matching
CUDA context. This prevents bugs when dealing with multiple CUDA contexts and
switching between these. Previously, it was assume that there was only a single
CUDA context.